move methods to helpers mod
