+++ smart-contracts/contracts/cl-vault/src/vault/range.rs
@@ -1,18 +1,16 @@
 use crate::{
     helpers::{
-        generic::extract_attribute_value_by_ty_and_key,
-        getters::{
+        assert::assert_range_admin, generic::extract_attribute_value_by_ty_and_key, getters::{
             get_single_sided_deposit_0_to_1_swap_amount,
             get_single_sided_deposit_1_to_0_swap_amount, get_twap_price, get_unused_balances,
-        },
-        msgs::swap_msg,
+        }, msgs::swap_msg
     },
     math::tick::price_to_tick,
     msg::{ExecuteMsg, MergePositionMsg},
     reply::Replies,
     state::{
         ModifyRangeState, Position, SwapDepositMergeState, CURRENT_BALANCE, CURRENT_SWAP,
-        MODIFY_RANGE_STATE, POOL_CONFIG, POSITION, RANGE_ADMIN, SWAP_DEPOSIT_MERGE_STATE,
+        MODIFY_RANGE_STATE, POOL_CONFIG, POSITION, SWAP_DEPOSIT_MERGE_STATE,
     },
     vault::{
         concentrated_liquidity::{create_position, get_position},
@@ -22,8 +20,8 @@ use crate::{
 };
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{
-    attr, to_json_binary, Addr, Coin, Decimal, Decimal256, Deps, DepsMut, Env, Fraction,
-    MessageInfo, Response, Storage, SubMsg, SubMsgResult, Uint128,
+    attr, to_json_binary, Coin, Decimal, Decimal256, DepsMut, Env, Fraction,
+    MessageInfo, Response, SubMsg, SubMsgResult, Uint128,
 };
 use osmosis_std::types::osmosis::{
     concentratedliquidity::v1beta1::{
@@ -38,18 +36,6 @@ use super::{
     concentrated_liquidity::get_cl_pool_info,
     swap::{SwapCalculationResult, SwapParams},
 };
-
-pub fn assert_range_admin(storage: &mut dyn Storage, sender: &Addr) -> Result<(), ContractError> {
-    let admin = RANGE_ADMIN.load(storage)?;
-    if admin != sender {
-        return Err(ContractError::Unauthorized {});
-    }
-    Ok(())
-}
-
-pub fn get_range_admin(deps: Deps) -> Result<Addr, ContractError> {
-    Ok(RANGE_ADMIN.load(deps.storage)?)
-}
 
 /// This function is the entrypoint into the dsm routine that will go through the following steps
 /// * how much liq do we have in current range
@@ -687,9 +673,7 @@ mod tests {
     use osmosis_std::types::osmosis::concentratedliquidity::v1beta1::MsgWithdrawPositionResponse;
 
     use crate::{
-        math::tick::build_tick_exp_cache,
-        state::{MODIFY_RANGE_STATE, RANGE_ADMIN},
-        test_helpers::{mock_deps_with_querier, mock_deps_with_querier_with_balance},
+        helpers::getters::get_range_admin, math::tick::build_tick_exp_cache, state::{MODIFY_RANGE_STATE, RANGE_ADMIN}, test_helpers::{mock_deps_with_querier, mock_deps_with_querier_with_balance}
     };
 
     #[test]
@@ -717,7 +701,7 @@ mod tests {
 
         RANGE_ADMIN.save(&mut deps.storage, &info.sender).unwrap();
 
-        assert_eq!(super::get_range_admin(deps.as_ref()).unwrap(), info.sender);
+        assert_eq!(get_range_admin(deps.as_ref()).unwrap(), info.sender);
     }
 
     #[test]
